Resource Usage: Do not error on missing data when fetching current
After fetching range, we error when there's no data matching the next
query.

The query fetches data for the timestamp `To` returned from the range
query. But the query has `request > usage`, so when that doesn't match
the query might return no rows.

In that case, for now, let's return empty data