Utilization: When fetching specific data for team, group rows usable for cost
When calculating trends and utilization, we want to include the overuse.
This PR switches to return up to two rows when requesting data which
previusly discarded in the case where usage was higher than request.

The code is not pretty, but seems to provide better values in the case
mentioned in #24

Fixes #24