feat: Update sorting logic in teams.resolvers.go
This commit updates the sorting logic in the `teams.resolvers.go` file. It adds cases for sorting by severity and risk score in the `Naisjobs` function, ensuring that the `naisjobs` slice is sorted correctly based on the specified order. This change improves the functionality of the GraphQL API by allowing users to retrieve Naisjobs in a sorted manner.

Co-authored-by: Roger Bjørnstad <roger.bjornstad@nav.no>
Co-authored-by: ybelMekk <youssef.bel.mekki@nav.no>