Enhance image resolver in app.resolvers.go and naisjob.resolvers.go
This commit enhances the image resolver in the app.resolvers.go and naisjob.resolvers.go files. It adds a loop to iterate through the image's workload references and fetches additional information about the referenced apps or jobs. This change improves the resolver's functionality by populating the DeployInfo field of the workload references, providing more comprehensive data for the GraphQL API.

Co-authored-by: Roger Bjørnstad <roger.bjornstad@nav.no>
Co-authored-by: ybelMekk <youssef.bel.mekki@nav.no>