feat: Refactor VulnerabilitiesSummary resolver in teams.resolvers.go
This commit refactors the VulnerabilitiesSummary resolver in the teams.resolvers.go file. It replaces the usage of the k8sClient to fetch apps with the dependencyTrackClient to get metadata for the team's slug. This change improves the resolver's efficiency by reducing unnecessary API calls and provides more accurate vulnerability summary data for the GraphQL API.

Co-authored-by: Roger Bjørnstad <roger.bjornstad@nav.no>
Co-authored-by: ybelMekk <youssef.bel.mekki@nav.no>