Don't add Get<FieldName> functions in generated interfaces
We have some interfaces in GraphQL, which will get Get<FieldName>
functions added to them by gqlgen. This gets somewhat confusing if the
field name has a resolver. The getters is not used by gqlgen, and is
currently not used by us either, so we can simply remove them, and stop
them from being generated.