feat: Add OpenSearchMetrics to OpenSearchList
This commit adds the OpenSearchMetrics field to the OpenSearchList type in the GraphQL schema. The OpenSearchMetrics field includes the cost of the OpenSearch instances. This allows for presenting the accumulated cost for a given cost type. The OpenSearch resolver in teams.resolvers.go has been updated to include the metrics in the OpenSearchList response.