feat: Add BucketsMetrics to BucketsList
This commit adds the BucketsMetrics field to the BucketsList type in the GraphQL schema. The BucketsMetrics field includes the cost of the buckets. This allows for presenting the accumulated cost for the buckets. The Buckets resolver in teams.resolvers.go has been updated to include the metrics in the BucketsList response.