Database name does not match SQL instance name
Instead of trying to get the database resource based on the SQL instance
name we need to list all SQL instances, and return the first database
instance that has a matching SQL instance name stored in its spec.

In theory multiple SQL databases with the same SQL instance reference
may exist in the cluster, but this is not supported in Nais, so we will
simply return the first match, if any.